+++ ratis-client/src/main/java/org/apache/ratis/client/impl/OrderedAsync.java
@@ -127,7 +127,7 @@ public final class OrderedAsync {
 
   private OrderedAsync(RaftClientImpl client, RaftProperties properties) {
     this.client = Objects.requireNonNull(client, "client == null");
-    this.requestSemaphore = new Semaphore(RaftClientConfigKeys.Async.outstandingRequestsMax(properties));
+    this.requestSemaphore = new Semaphore(RaftClientConfigKeys.DataStream.outstandingRequestsMax(properties));
   }
 
   private void resetSlidingWindow(RaftClientRequest request) {
