+++ ratis-common/src/main/java/org/apache/ratis/datastream/impl/DataStreamReplyByteBuffer.java
@@ -15,7 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ratis.protocol;
+package org.apache.ratis.datastream.impl;
+
+import org.apache.ratis.protocol.DataStreamReply;
 
 import java.nio.ByteBuffer;
