+++ ratis-netty/src/main/java/org/apache/ratis/netty/server/NettyServerStreamRpc.java
@@ -20,7 +20,7 @@ package org.apache.ratis.netty.server;
 
 import org.apache.ratis.client.impl.ClientProtoUtils;
 import org.apache.ratis.proto.RaftProtos;
-import org.apache.ratis.protocol.DataStreamReplyByteBuffer;
+import org.apache.ratis.datastream.impl.DataStreamReplyByteBuffer;
 import org.apache.ratis.protocol.RaftClientRequest;
 import org.apache.ratis.protocol.RaftPeer;
 import org.apache.ratis.server.DataStreamServerRpc;
@@ -88,22 +88,22 @@ public class NettyServerStreamRpc implements DataStreamServerRpc {
       }
 
       final WritableByteChannel channel = stream.getWritableByteChannel();
-      long written = 0;
+      long byteWritten = 0;
       for (ByteBuffer buffer : buf.nioBuffers()) {
         try {
-          written += channel.write(buffer);
+          byteWritten += channel.write(buffer);
         } catch (Throwable t) {
           throw new CompletionException(t);
         }
       }
-      return written;
+      return byteWritten;
     } finally {
       buf.release();
     }
   }
 
-  private void sendReply(DataStreamRequestByteBuf request, long written, ChannelHandlerContext ctx) {
-    final boolean success = request.getDataLength() == written;
+  private void sendReply(DataStreamRequestByteBuf request, long byteWritten, ChannelHandlerContext ctx) {
+    // TODO RATIS-1098: include byteWritten and isSuccess in the reply
     final DataStreamReplyByteBuffer reply = new DataStreamReplyByteBuffer(
         request.getStreamId(), request.getStreamOffset(), ByteBuffer.wrap("OK".getBytes()));
     ctx.writeAndFlush(reply);
@@ -113,13 +113,12 @@ public class NettyServerStreamRpc implements DataStreamServerRpc {
     return new ChannelInboundHandlerAdapter(){
       @Override
       public void channelRead(ChannelHandlerContext ctx, Object msg) {
-        final DataStreamRequestByteBuf req = (DataStreamRequestByteBuf)msg;
-        final long streamId = req.getStreamId();
-        final ByteBuf buf = req.slice();
+        final DataStreamRequestByteBuf request = (DataStreamRequestByteBuf)msg;
+        final ByteBuf buf = request.slice();
         final AtomicBoolean released = new AtomicBoolean();
-        streams.computeIfAbsent(streamId, id -> getDataStreamFuture(buf, released))
+        streams.computeIfAbsent(request.getStreamId(), id -> getDataStreamFuture(buf, released))
             .thenApply(stream -> writeTo(buf, stream, released.get()))
-            .thenAccept(written -> sendReply(req, written, ctx));
+            .thenAccept(byteWritten -> sendReply(request, byteWritten, ctx));
       }
     };
   }
